Refactoring, code cleaning, exception handling and access handling divided into two different classes-a
