+++ src/main/java/com/evolveum/polygon/scim/HandlingStrategy.java
@@ -7,7 +7,10 @@ import java.util.Set;
 import org.apache.http.Header;
 import org.apache.http.HttpResponse;
 import org.apache.http.message.BasicHeader;
+import org.identityconnectors.framework.common.exceptions.ConnectionFailedException;
 import org.identityconnectors.framework.common.exceptions.ConnectorException;
+import org.identityconnectors.framework.common.exceptions.ConnectorIOException;
+import org.identityconnectors.framework.common.exceptions.UnknownUidException;
 import org.identityconnectors.framework.common.objects.Attribute;
 import org.identityconnectors.framework.common.objects.AttributeInfoBuilder;
 import org.identityconnectors.framework.common.objects.ConnectorObject;
@@ -44,15 +47,105 @@ public interface HandlingStrategy {
 
 	Header PRETTYPRINTHEADER = new BasicHeader("X-PrettyPrint", "1");
 
+	/**
+	 * Sends queries for object creation to the service providers endpoints.
+	 * After successful object creation the service provider returns the uid of
+	 * the created object.
+	 * 
+	 * @param objectTranslator
+	 *            An instance of object translator containing methods for the
+	 *            creation of an json object out of an provided set of
+	 *            attributes.
+	 * @param resourceEndPoint
+	 *            The resource endpoint name.
+	 * @param attributes
+	 *            The provided attributes set containing information for object
+	 *            creation.
+	 * @throws ConnectorException
+	 * @throws ConnectorIOException
+	 * @throws ConnectionFailedException
+	 * @throws UnknownUidException
+	 *
+	 * @return the uid of the created object.
+	 */
+
 	public Uid create(String resourceEndPoint, ObjectTranslator objectTranslator, Set<Attribute> attributes,
 			Set<Attribute> injectedAttributeSet, ScimConnectorConfiguration conf);
+
+	/**
+	 * Sends queries to the service provider endpoints to retrieve the queried
+	 * information and processes responses which are handed over to the provided
+	 * result handler.
+	 * 
+	 * @param query
+	 *            The query object which can be a string or an Uid type object.
+	 * @param resourceEndPoint
+	 *            The resource endpoint name.
+	 * @param resultHandler
+	 *            The provided result handler which handles results.
+	 *
+	 * @throws ConnectorException
+	 * @throws ConnectorIOException
+	 */
 
 	public void qeuery(Object query, String resourceEndPoint, ResultsHandler resultHandler,
 			ScimConnectorConfiguration conf);
 
+	/**
+	 * Sends queries for object update to the service providers endpoints. After
+	 * successful object update the service provider returns the uid of the
+	 * updated object.
+	 * 
+	 * @param uid
+	 *            The uid of the resource object which should be updated.
+	 * @param resourceEndPoint
+	 *            The name of the resource endpoint.
+	 * @param jsonObject
+	 *            The json object which carries the information which should be
+	 *            updated.
+	 * 
+	 * @throws ConnectorException
+	 * @throws UnknownUidException
+	 * @throws ConnectorIOException
+	 * @throws ConnectionFailedException
+	 *
+	 * @return the uid of the created object.
+	 */
+
 	public Uid update(Uid uid, String resourceEndPoint, JSONObject jsonObject, ScimConnectorConfiguration conf);
 
+	/**
+	 * Sends queries for object deletion to the service providers endpoints.
+	 * 
+	 * @param uid
+	 *            The uid of the resource object which should be deleted.
+	 * @param resourceEndPoint
+	 *            The name of the resource endpoint.
+	 * 
+	 * @throws ConnectorIOException
+	 * @throws ConnectionFailedException
+	 *
+	 */
+
 	public void delete(Uid uid, String resourceEndPoint, ScimConnectorConfiguration conf);
+
+	/**
+	 * Sends queries to the service provider endpoints to retrieve the queried
+	 * information and processes responses which are handed over to the provided
+	 * result handler. This method send queries only to the endpoints containing
+	 * the schema information of the services resources.
+	 * 
+	 * @param providerName
+	 *            The name of the provider.
+	 * @param resourceEndPoint
+	 *            The resource endpoint name.
+	 *
+	 * @throws ConnectorException
+	 * @throws ConnectorIOException
+	 *
+	 * @return an instance of "ScimSchemaParser" containing the schema
+	 *         information of all endpoint.
+	 */
 
 	public ParserSchemaScim qeuerySchemas(String providerName, String resourceEndPoint,
 			ScimConnectorConfiguration conf);
@@ -61,7 +154,7 @@ public interface HandlingStrategy {
 
 	public ParserSchemaScim processSchemaResponse(JSONObject responseObject);
 
-	public Map<String, Map<String, Object>> parseAttribute(JSONObject attribute,
+	public Map<String, Map<String, Object>> parseSchemaAttribute(JSONObject attribute,
 			Map<String, Map<String, Object>> attributeMap, ParserSchemaScim parser);
 
 	public List<Map<String, Map<String, Object>>> getAttributeMapList(
@@ -70,7 +163,7 @@ public interface HandlingStrategy {
 	public Map<String, Object> translateReferenceValues(Map<String, Map<String, Object>> attributeMap,
 			JSONArray referenceValues, Map<String, Object> subAttributeMap, int position, String attributeName);
 
-	public Set<Attribute> addAttributeToInject(Set<Attribute> injectetAttributeSet);
+	public Set<Attribute> addAttributesToInject(Set<Attribute> injectetAttributeSet);
 
 	public Uid groupUpdateProcedure(HttpResponse response, JSONObject jsonObject, String uri, Header authHeader);
